Rename test module from rpc_tests to tests
Consolidates all RPC tests into a single external file following CLAUDE.md
convention: "#[path = "file.rs"] #[cfg(test)] mod tests;"

- Moved inline cache tests from rpc.rs to rpc_tests.rs
- Renamed module from `rpc_tests` to `tests` while keeping file as rpc_tests.rs
- All 9 tests pass (4 serialization + 5 cache tests)

Per review feedback: "Test module should be named tests, not rpc_tests."
